Criação e finalização do CreditHandler com Testes

+++ PagueVeloz.Core/Application/DTOs/Transaction/TransactionResponse.cs
@@ -4,9 +4,9 @@
     {
         public string TransactionId { get; set; } = string.Empty;
         public string Status { get; set; } = "pending"; // success, failed, pending
-        public decimal Balance { get; set; }
-        public decimal ReservedBalance { get; set; }
-        public decimal AvailableBalance { get; set; }
+        public long Balance { get; set; }
+        public long ReservedBalance { get; set; }
+        public long AvailableBalance { get; set; }
         public DateTime Timestamp { get; set; } = DateTime.UtcNow;
         public string? ErrorMessage { get; set; }
     }
